// src/store/modules/upload/mutation-types.ts
export enum MutationTypes {
  UPLOAD_CONFIGURATIONS_LOADING = "UPLOAD_CONFIGURATIONS_LOADING",
  UPLOAD_CONFIGURATIONS_LOADED = "UPLOAD_CONFIGURATIONS_LOADED",
  UPLOAD_FILE_LOADING = "UPLOAD_FILE_LOADING",
  UPLOAD_FILE_LOADED = "UPLOAD_FILE_LOADED",
  UPLOADED_FILES_LOADING = "UPLOADED_FILES_LOADING",
  UPLOADED_FILES_LOADED = "UPLOADED_FILES_LOADED",
  DELETE_FILE_LOADING = "DELETE_FILE_LOADING",
  DELETE_FILE_LOADED = "DELETE_FILE_LOADED",
}